Split QueryGraph
